requirements.txt
requests
notion-client
feedparser
python-dotenv
virustotal-python
OTXv2
beautifulsoup4
